pyproject.toml
[project]
name = "coding-everything"
version = "0.2.0"
description = "Personal AI coding assistant configuration manager"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
ce = "install_skills.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["install_skills"]
